requirements.txt
httpx==0.27.2
selectolax==0.3.26
pydantic==2.10.3
sqlmodel==0.0.16
python-dotenv==1.0.1
backoff==2.2.1
typer==0.12.5
structlog==24.4.0
rich==13.9.4
python-telegram-bot==22.1.0
gspread-pandas==3.3.0
pandas==2.2.3
anyio==4.6.2
aiosqlite==0.20.0
firecrawl-py==2.15.0
pydantic-settings==2.5.2
# Additional dependencies for core data layer
sqlalchemy[asyncio]==2.0.23
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.8.0
ruff==0.6.9
# Web scraping dependencies
beautifulsoup4==4.12.3
lxml==5.3.0
# Configuration file support
pyyaml==6.0.2